Only do the initial fade-out on keyup inside of TinyMCE.
This means fewer fade-outs when you are just absentmindedly
passing through the editor or using the mouse to select something
to bold or link. But when you start typing, and your focus is
assuredly on the content, you get the fade out. The fade-out on
mouse return while focus remains is still in place.

The title is now not a fade-out trigger. You don't need to be
distraction-free while writing a title, and titles are often
quickly edited or stubbed out, so the fade/return cycle is a
bit distracting there.

I also made the return speed faster.

fixes #28